Added BM25 to DAAT

+++ index/query_processing/ranking_functions/BM25.hpp
@@ -4,6 +4,6 @@
 #include <vector>
 
 //freq and docscontaining indexes correspond to the same term
-double BM25(std::vector<unsigned int>& freq, std::vector<unsigned int>& docscontaining, double doclength, double avgdoclength, double totaldocs);
+double BM25(std::vector<unsigned int>& freq, std::vector<unsigned int>& docscontaining, unsigned int doclength, double avgdoclength, unsigned int totaldocs);
 
 #endif
